+++ crates/fluvio-smartengine/src/smartmodule/aggregate.rs
@@ -7,6 +7,7 @@ use wasmtime::{AsContextMut, Trap, TypedFunc};
 use crate::{
     WasmSlice,
     smartmodule::{SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance},
+    error::Error,
 };
 use dataplane::smartmodule::{
     SmartModuleAggregateInput, SmartModuleInput, SmartModuleOutput, SmartModuleInternalError,
@@ -42,7 +43,7 @@ impl SmartModuleAggregate {
         params: SmartModuleExtraParams,
         accumulator: Vec<u8>,
         version: i16,
-    ) -> Result<Self> {
+    ) -> Result<Self, Error> {
         let mut base = SmartModuleContext::new(module, params, version)?;
         let aggregate_fn: AggregateFnKind = if let Ok(agg_fn) = base
             .instance
@@ -52,7 +53,8 @@ impl SmartModuleAggregate {
         } else {
             let agg_fn: OldAggregateFn = base
                 .instance
-                .get_typed_func(&mut base.store, AGGREGATE_FN_NAME)?;
+                .get_typed_func(&mut base.store, AGGREGATE_FN_NAME)
+                .map_err(|err| Error::NotNamedExport(AGGREGATE_FN_NAME, err))?;
             AggregateFnKind::Old(agg_fn)
         };